# docs/api/storage_openapi.yaml
openapi: 3.0.1
info:
  title: Storage Service
  description: Storage service which handles the metadata ingestion in the OSDU data platform
  contact:
    name: OSDU Data Platform Team
    email: dps@OSDU.org
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 2.0.0
servers:
  - url: /api/storage/v2/
security:
  - Authorization: []
...
paths:
  /records:
    get:
      tags:
        - records
      summary: Fetch All records
      description: |-
        This API retrieves multiple records in a single request with support for filtering by kind, deleted status and lastModifiedDate. It is paginated, using a cursor parameter to fetch subsequent pages.The default page size is 20, configurable via the limit parameter.
        Allowed roles: `service.storage.viewer`,`service.storage.creator` and `service.storage.admin`.
      operationId: getAllRecords
      parameters:
        - name: x-collaboration
          in: header
          description: x-collaboration
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description: Page Size
          required: false
          schema:
            maximum: 100
            minimum: 1
            type: integer
            format: int32
            default: 20
          example: 20
        - name: kind
          in: query
          description: Filter Kind
          required: false
          schema:
            pattern: ^[\w\-\.]+:[\w\-\.]+:[\w\-\.]+:[0-9]+.[0-9]+.[0-9]+$
            type: string
          example: tenant1:public:well:1.0.2
        - name: cursor
          in: query
          description: Cursor
          required: false
          schema:
            type: string
        - name: deleted
          in: query
          description: Get Only soft deleted records in response
          required: false
          schema:
            type: boolean
            default: false
          example: true
        - name: modifiedAfterDate
          in: query
          description: Get records modified after this date
          required: false
          schema:
            type: string
            format: date-time
          example: '2025-09-04'
        - name: sortOrder
          in: query
          description: Sort Order
          required: false
          schema:
            type: string
            default: DESC
            enum:
              - ASC
              - DESC
        - name: data-partition-id
          in: header
          description: Tenant Id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Record retrieved successfully.
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '404':
          description: Record not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '502':
          description: Bad Gateway
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
      security:
        - Authorization: []
    put:
      tags:
        - records
      summary: Create or Update Records
      description: |-
        The API represents the main injection mechanism into the Data Ecosystem. 
        It allows records creation and/or update.When no record id is provided or when the provided id is not already present in the Data Ecosystem then a new record is created. 
         If the id is related to an existing record in the Data Ecosystem then an update operation takes place and a new version of the record is created.
      operationId: createOrUpdateRecords
      parameters:
        - name: x-collaboration
          in: header
          description: x-collaboration
          required: false
          schema:
            type: string
        - name: skipdupes
          in: query
          description: Skip duplicates when updating records with the same value.
          required: false
          schema:
            type: boolean
        - name: data-partition-id
          in: header
          description: Tenant Id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              maxItems: 500
              minItems: 0
              type: array
              description: Records to be created/updated
              items:
                $ref: '#/components/schemas/Record'
        required: true
      responses:
        '201':
          description: Records created and/or updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUpdateRecordsResponse'
        '400':
          description: Invalid record format.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '403':
          description: User not authorized to perform the action.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '404':
          description: Invalid acl group.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '502':
          description: Bad Gateway
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
      security:
        - Authorization: []
    patch:
      tags:
        - records
      summary: Modify record data and/or metadata attributes using patch operations
      description: |-
        The API represents the patch update mechanism for records. It allows updating multiple records in one request. The API supports metadata update only (Legal Tags, ACLs and Tags) if the request body media type is `application/json`. The API supports metadata and data update (Legal Tags, ACLs, Tags, Ancestry, Kind, Meta and Data) if the request body media type is `application/json-patch+json`. Please choose the appropriate media type from the Request body dropdown. The currently supported operations are replace, add, and remove. 
        Required roles: `users.datalake.editors` or `users.datalake.admins`.
      operationId: updateRecordsMetadata_1
      parameters:
        - name: x-collaboration
          in: header
          description: x-collaboration
          required: false
          schema:
            type: string
        - name: data-partition-id
          in: header
          description: Tenant Id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordBulkUpdateParam'
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/PatchRecordsRequestModel'
        required: true
      responses:
        '200':
          description: Records updated successfully.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/PatchRecordsResponse'
                  - $ref: '#/components/schemas/BulkUpdateRecordsResponse'
        '206':
          description: Records updated successful partially.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/PatchRecordsResponse'
                  - $ref: '#/components/schemas/BulkUpdateRecordsResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '403':
          description: User not authorized to perform the action.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '502':
          description: Bad Gateway
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
      security:
        - Authorization: []
  /records/copy:
    put:
      tags:
        - records
      summary: Copy Record references form one namespace to another
      description: This API attempts to copy all the Record references it is provided from the given source namespace to the target namespace. All references will be copied or all will fail as a transaction. If the target namespace does not et exist it will be created. It requires 'services.storage.admin' permission to call
      operationId: copyRecordReferencesBetweenNamespaces
      parameters:
        - name: x-collaboration
          in: header
          description: x-collaboration
          required: true
          schema:
            type: string
        - name: data-partition-id
          in: header
          description: Tenant Id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CopyRecordReferencesModel'
        required: true
      responses:
        '200':
          description: Record references copied successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CopyRecordReferencesModel'
        '400':
          description: Invalid record ids provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '403':
          description: User not authorized to perform the action.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '404':
          description: Records not found in the source.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '409':
          description: One or more references already exist in the target namespace.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '502':
          description: Bad Gateway
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
      security:
        - Authorization: []
...
components:
  schemas:
    Acl:
      type: object
      properties:
        viewers:
          type: array
          items:
            type: string
        owners:
          type: array
          items:
            type: string
    Legal:
      required:
        - otherRelevantDataCountries
      type: object
      properties:
        legaltags:
          uniqueItems: true
          type: array
          items:
            type: string
        otherRelevantDataCountries:
          uniqueItems: true
          type: array
          items:
            type: string
        status:
          type: string
          enum:
            - compliant
            - incompliant
    Record:
      required:
        - acl
        - data
      type: object
      properties:
        id:
          pattern: ^[\w\-\.]+:[\w-\.]+:[\w\-\.\:\%]+$
          type: string
        version:
          type: integer
          format: int64
        kind:
          type: string
        acl:
          $ref: '#/components/schemas/Acl'
        legal:
          $ref: '#/components/schemas/Legal'
        data:
          type: object
          additionalProperties:
            type: object
        ancestry:
          $ref: '#/components/schemas/RecordAncestry'
        meta:
          type: array
          items:
            type: object
            additionalProperties:
              type: object
        tags:
          type: object
          additionalProperties:
            type: string
        createUser:
          type: string
        createTime:
          type: string
        modifyUser:
          type: string
        modifyTime:
          type: string
    RecordAncestry:
      type: object
      properties:
        parents:
          uniqueItems: true
          type: array
          items:
            type: string
    AppError:
      type: object
      properties:
        code:
          type: integer
          format: int32
        reason:
          type: string
        message:
          type: string
    CreateUpdateRecordsResponse:
      type: object
      properties:
        recordCount:
          type: integer
          format: int32
        recordIds:
          type: array
          items:
            type: string
        skippedRecordIds:
          type: array
          items:
            type: string
        recordIdVersions:
          type: array
          items:
            type: string
    CopyRecordReferencesModel:
      type: object
      properties:
        target:
          type: string
        records:
          type: array
          items:
            $ref: '#/components/schemas/RecordVersionModel'
    RecordVersionModel:
      type: object
      properties:
        id:
          type: string
        version:
          type: string
...
    PatchOperation:
      required:
        - value
      type: object
      properties:
        op:
          type: string
        path:
          type: string
        value:
          type: array
          items:
            type: string
    RecordBulkUpdateParam:
      required:
        - ops
        - query
      type: object
      properties:
        query:
          $ref: '#/components/schemas/RecordQuery'
        ops:
          type: array
          items:
            $ref: '#/components/schemas/PatchOperation'
      description: Records to be updated
    RecordQuery:
      required:
        - ids
      type: object
      properties:
        ids:
          maxItems: 500
          minItems: 1
          type: array
          items:
            type: string
    BulkUpdateRecordsResponse:
      type: object
      properties:
        recordCount:
          type: integer
          format: int32
        recordIds:
          type: array
          items:
            type: string
        notFoundRecordIds:
          type: array
          items:
            type: string
        unAuthorizedRecordIds:
          type: array
          items:
            type: string
        lockedRecordIds:
          type: array
          items:
            type: string
    JsonPatch:
      type: object
      description: Record patch operations.
    PatchRecordsRequestModel:
      required:
        - ops
        - query
      type: object
      properties:
        query:
          $ref: '#/components/schemas/RecordQueryPatch'
        ops:
          $ref: '#/components/schemas/JsonPatch'
      description: Records to be patched
      example:
        query:
          ids:
            - common:work-product-component--wellLog:123456
        ops:
          - op: remove
            path: /acl/viewers/0
    RecordQueryPatch:
      type: object
      properties:
        ids:
          type: array
          description: List of record ids. Each record id must follow the naming convention {Data-Partition-Id}:{dataset-name}:{record-type}:{version}.
          items:
            type: string
      description: Represents a model for Record Query Patch
      example:
        ids:
          - common:work-product-component--wellLog:123456
    PatchRecordsResponse:
      type: object
      properties:
        recordCount:
          type: integer
          format: int32
        recordIds:
          type: array
          items:
            type: string
        notFoundRecordIds:
          type: array
          items:
            type: string
        failedRecordIds:
          type: array
          items:
            type: string
        errors:
          type: array
          items:
            type: string
...
  securitySchemes:
    Authorization:
      type: http
      scheme: bearer
      bearerFormat: Authorization